+++ packages/svelte-reveal/src/internal/validations.ts
@@ -1,4 +1,5 @@
 import { ROOT_MARGIN_REGEX } from './constants.ts';
+import { Responsive } from './types/devices.ts';
 import type { RevealOptions } from './types/options.ts';
 
 /**
@@ -26,6 +27,35 @@ export const isPositive = (value: number) => value >= 0;
  */
 export const isPositiveInteger = (value: number) => {
   return isPositive(value) && Number.isInteger(value);
+};
+
+/**
+ * Checks whether the breakpoints overlap.
+ * @param responsive An object that instructs the library how to handle responsiveness for a given set of devices.
+ * @returns Whether the breapoints overlap.
+ */
+export const hasOverlappingBreakpoints = (responsive: Responsive): boolean => {
+  const { mobile, tablet, laptop, desktop } = responsive;
+
+  return (
+    mobile.breakpoint > tablet.breakpoint ||
+    tablet.breakpoint > laptop.breakpoint ||
+    laptop.breakpoint > desktop.breakpoint
+  );
+};
+
+/**
+ * Checks whether the breakpoints are valid or not.
+ * @param responsive An object that instructs the library how to handle responsiveness for a given set of devices.
+ * @returns Whether the breakpoints are valid.
+ */
+export const hasValidBreakpoints = (responsive: Responsive): boolean => {
+  const breakpoints = Object.values(responsive).map((device) => device.breakpoint);
+
+  const doBreakpointsOverlap = hasOverlappingBreakpoints(responsive);
+  const allBreakpointsPositive = breakpoints.every((breakpoint) => isPositiveInteger(breakpoint));
+
+  return !doBreakpointsOverlap && allBreakpointsPositive;
 };
 
 /**
